+++ packages/compiler/src/parser.ts
@@ -628,10 +628,8 @@ export const parse = (tokens: (Token | undefined)[]) => {
       try {
         switch( keyword ) {
           case 'owned': {
-            if( match(TokenTypes.QuotedString, 'on-write') ) {
-              node.owned = consume(TokenTypes.QuotedString).value === 'on-write'
-            } else {
-              node.owned = consume(TokenTypes.Boolean).value
+            if( match(TokenTypes.QuotedString, ['always', 'on-write']) ) {
+              node.owned = consume(TokenTypes.QuotedString).value as AST.CollectionNode['owned']
             }
             break
           }