+++ packages/http/src/options.ts
@@ -1,9 +1,0 @@
-export type ServerOptions = {
-  host?: string
-  port?: number
-}
-
-export const defineServerOptions = <TServerOptions extends ServerOptions>(options: TServerOptions) => {
-  return options
-}
-